토큰시간 표시
