+++ src/main/moivo_react/src/assets/css/admins_qnaboard.module.css
@@ -518,26 +518,3 @@
     margin-bottom: 30px;
     z-index: 100;
 }
-
-/*  토큰 시간  */
-.tokenTimerContainer {
-    position: fixed;
-    top: 20px;
-    right: 20px;
-    z-index: 1000;
-}
-
-.tokenTimer {
-    background-color: rgba(0, 0, 0, 0.7);
-    color: white;
-    padding: 8px 15px;
-    border-radius: 20px;
-    font-size: 14px;
-    display: flex;
-    align-items: center;
-    gap: 8px;
-}
-
-.tokenTimer i {
-    color: #ffd700;
-}
